Don't remove trailing "_list" in data types.
Since we're moving to standard JSON types, the existing data types are
renamed. These were sometimes using a suffix "_list" to indicate that
they're actually a comma separated list. It is used by the rsmp_schema.
However, we were omitting this suffix in the spec.

Now that we're about to add JSON data types and the legacy types are
renamed, this can be removed.